Car-specific controls, championship competition message

+++ AcManager.Tools/GameProperties/WeatherSpecificCloudsHelper.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
+using AcManager.Tools.Helpers.AcSettings;
 using AcManager.Tools.Managers;
 using AcManager.Tools.Objects;
 using AcTools.Utils;
 using FirstFloor.ModernUI.Helpers;
+using JetBrains.Annotations;
 
 namespace AcManager.Tools.GameProperties {
     internal interface IWeatherSpecificReplacement {
@@ -13,27 +15,23 @@ namespace AcManager.Tools.GameProperties {
         void Revert();
     }
 
-    internal class WeatherSpecificDirectoryReplacement : IWeatherSpecificReplacement {
-        public string RelativeSource { get; }
+    internal abstract class TemporaryDirectoryReplacementBase {
+        private readonly string _relativeDestination;
+        private readonly string _relativeBackup;
 
-        public string RelativeDestination { get; }
-
-        protected string RelativeBackup { get; }
-
-        internal WeatherSpecificDirectoryReplacement(string relativeRelativeSource, string relativeDestination) {
-            RelativeSource = relativeRelativeSource;
-            RelativeDestination = relativeDestination;
-            RelativeBackup = RelativeDestination + @"_backup_cm";
+        internal TemporaryDirectoryReplacementBase([NotNull] string relativeDestination, string backupPostfix = @"_backup_cm") {
+            _relativeDestination = relativeDestination;
+            _relativeBackup = _relativeDestination + backupPostfix;
         }
 
-        public bool Apply(WeatherObject weather) {
-            if (AcRootDirectory.Instance.Value == null) return false;
+        [NotNull]
+        protected abstract string GetAbsolutePath([NotNull] string relative);
 
-            var source = Path.Combine(weather.Location, RelativeSource);
-            if (!Directory.Exists(source)) return false;
+        protected bool Apply([NotNull] string source) {
+            if (AcRootDirectory.Instance.Value == null || !Directory.Exists(source)) return false;
 
-            var destination = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeDestination);
-            var backup = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeBackup);
+            var destination = GetAbsolutePath(_relativeDestination);
+            var backup = GetAbsolutePath(_relativeBackup);
 
             if (Directory.Exists(destination)) {
                 if (Directory.Exists(backup)) {
@@ -44,14 +42,13 @@ namespace AcManager.Tools.GameProperties {
                 Directory.Move(destination, backup);
             }
 
-
             try {
                 Logging.Debug($"{source} → {destination}");
                 FileUtils.CopyRecursiveHardlink(source, destination);
             } catch (Exception e) {
                 // this exception should be catched here so original clouds folder still
                 // will be restored even when copying a new one has been failed
-                NonfatalError.Notify("Can’t replace weather-specific directory", e);
+                NonfatalError.Notify("Can’t replace directory", e);
             }
 
             return true;
@@ -60,8 +57,8 @@ namespace AcManager.Tools.GameProperties {
         public void Revert() {
             if (AcRootDirectory.Instance.Value == null) return;
 
-            var destination = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeDestination);
-            var backup = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeBackup);
+            var destination = GetAbsolutePath(_relativeDestination);
+            var backup = GetAbsolutePath(_relativeBackup);
 
             try {
                 if (Directory.Exists(backup)) {
@@ -72,32 +69,28 @@ namespace AcManager.Tools.GameProperties {
                     Directory.Move(backup, destination);
                 }
             } catch (Exception e) {
-                NonfatalError.Notify("Can’t restore original directory after replacing it with a weather-specific one", e);
+                NonfatalError.Notify("Can’t restore original directory after replacing it with a temporary one", e);
             }
         }
     }
 
-    internal class WeatherSpecificFileReplacement : IWeatherSpecificReplacement {
-        public string RelativeSource { get; }
+    internal abstract class TemporaryFileReplacementBase {
+        private readonly string _relativeDestination;
+        private readonly string _relativeBackup;
 
-        public string RelativeDestination { get; }
-
-        protected string RelativeBackup { get; }
-
-        internal WeatherSpecificFileReplacement(string relativeRelativeSource, string relativeDestination) {
-            RelativeSource = relativeRelativeSource;
-            RelativeDestination = relativeDestination;
-            RelativeBackup = RelativeDestination + @"_backup_cm";
+        internal TemporaryFileReplacementBase([NotNull] string relativeDestination, string backupPostfix = @"_backup_cm") {
+            _relativeDestination = relativeDestination;
+            _relativeBackup = _relativeDestination + backupPostfix;
         }
 
-        public bool Apply(WeatherObject weather) {
-            if (AcRootDirectory.Instance.Value == null) return false;
+        [NotNull]
+        protected abstract string GetAbsolutePath([NotNull] string relative);
 
-            var source = Path.Combine(weather.Location, RelativeSource);
-            if (!File.Exists(source)) return false;
+        protected bool Apply([NotNull] string source) {
+            if (AcRootDirectory.Instance.Value == null || !File.Exists(source)) return false;
 
-            var destination = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeDestination);
-            var backup = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeBackup);
+            var destination = GetAbsolutePath(_relativeDestination);
+            var backup = GetAbsolutePath(_relativeBackup);
             if (File.Exists(destination)) {
                 if (File.Exists(backup)) {
                     File.Move(backup, FileUtils.EnsureUnique(backup));
@@ -113,7 +106,7 @@ namespace AcManager.Tools.GameProperties {
             } catch (Exception e) {
                 // this exception should be catched here so original clouds folder still
                 // will be restored even when copying a new one has been failed
-                NonfatalError.Notify("Can’t replace weather-specific files", e);
+                NonfatalError.Notify("Can’t replace files", e);
             }
 
             return true;
@@ -123,8 +116,8 @@ namespace AcManager.Tools.GameProperties {
             if (AcRootDirectory.Instance.Value == null) return;
 
             try {
-                var destination = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeDestination);
-                var backup = Path.Combine(AcRootDirectory.Instance.RequireValue, RelativeBackup);
+                var destination = GetAbsolutePath(_relativeDestination);
+                var backup = GetAbsolutePath(_relativeBackup);
 
                 if (File.Exists(backup)) {
                     if (File.Exists(destination)) {
@@ -134,14 +127,63 @@ namespace AcManager.Tools.GameProperties {
                     File.Move(backup, destination);
                 }
             } catch (Exception e) {
-                NonfatalError.Notify("Can’t restore original files after replacing them with weather-specific ones", e);
+                NonfatalError.Notify("Can’t restore original files after replacing them with temporary ones", e);
             }
+        }
+    }
+
+    internal class TemporaryFileReplacement : TemporaryFileReplacementBase {
+        private readonly string _source;
+
+        public TemporaryFileReplacement([CanBeNull] string source, [NotNull] string relativeDestination, string backupPostfix = @"_backup_cm")
+                : base(relativeDestination, backupPostfix) {
+            _source = source;
+        }
+
+        public bool Apply() {
+            return _source != null && Apply(_source);
+        }
+
+        protected override string GetAbsolutePath(string relative) {
+            return relative;
+        }
+    }
+
+    internal class WeatherSpecificDirectoryReplacementBase : TemporaryDirectoryReplacementBase, IWeatherSpecificReplacement {
+        public string RelativeSource { get; }
+
+        internal WeatherSpecificDirectoryReplacementBase(string relativeSource, string relativeDestination) : base(relativeDestination) {
+            RelativeSource = relativeSource;
+        }
+
+        public bool Apply(WeatherObject weather) {
+            return Apply(Path.Combine(weather.Location, RelativeSource));
+        }
+
+        protected override string GetAbsolutePath(string relative) {
+            return Path.Combine(AcRootDirectory.Instance.RequireValue, relative);
+        }
+    }
+
+    internal class WeatherSpecificFileReplacementBase : TemporaryFileReplacementBase, IWeatherSpecificReplacement {
+        public string RelativeSource { get; }
+
+        internal WeatherSpecificFileReplacementBase(string relativeRelativeSource, string relativeDestination) : base(relativeDestination) {
+            RelativeSource = relativeRelativeSource;
+        }
+
+        public bool Apply(WeatherObject weather) {
+            return Apply(Path.Combine(weather.Location, RelativeSource));
+        }
+
+        protected override string GetAbsolutePath(string relative) {
+            return Path.Combine(AcRootDirectory.Instance.RequireValue, relative);
         }
     }
 
     public class WeatherSpecificCloudsHelper : WeatherSpecificHelperBase {
         private static readonly IWeatherSpecificReplacement[] Replacements = {
-            new WeatherSpecificDirectoryReplacement(@"clouds", @"content\texture\clouds"),
+            new WeatherSpecificDirectoryReplacementBase(@"clouds", @"content\texture\clouds"),
         };
 
         public static void Revert() {
@@ -161,8 +203,8 @@ namespace AcManager.Tools.GameProperties {
 
     public class WeatherSpecificTyreSmokeHelper : WeatherSpecificHelperBase {
         private static readonly IWeatherSpecificReplacement[] Replacements = {
-            new WeatherSpecificFileReplacement(@"tyre_smoke.ini", @"system\cfg\tyre_smoke.ini"), 
-            new WeatherSpecificFileReplacement(@"tyre_smoke_grass.ini", @"system\cfg\tyre_smoke_grass.ini")
+            new WeatherSpecificFileReplacementBase(@"tyre_smoke.ini", @"system\cfg\tyre_smoke.ini"), 
+            new WeatherSpecificFileReplacementBase(@"tyre_smoke_grass.ini", @"system\cfg\tyre_smoke_grass.ini")
         };
 
         public static void Revert() {
@@ -173,6 +215,23 @@ namespace AcManager.Tools.GameProperties {
 
         protected override bool SetOverride(WeatherObject weather) {
             return Replacements.Aggregate(false, (current, replacement) => replacement.Apply(weather) || current);
+        }
+
+        protected override void DisposeOverride() {
+            Revert();
+        }
+    }
+
+    public class CarSpecificControlsPresetHelper : CarSpecificHelperBase {
+        private const string BackupPostfix = "_backup_cm_carspec";
+
+        public static void Revert() {
+            new TemporaryFileReplacement(null, AcSettingsHolder.Controls.Filename, BackupPostfix).Revert();
+        }
+
+        protected override bool SetOverride(CarObject car) {
+            return car.SpecificControlsPreset && car.ControlsPresetFilename != null &&
+                    new TemporaryFileReplacement(car.ControlsPresetFilename, AcSettingsHolder.Controls.Filename, BackupPostfix).Apply();
         }
 
         protected override void DisposeOverride() {
